Fix compilation and urgent issues
2.0.0 now compiles

Fixes included in this commit:
- Conditions work again
- Placeholders work again
- Leaving removes DisplayUser objects
- The player can join the server again

+++ core/src/main/java/better/scoreboard/core/animation/impl/HealthAnimation.java
@@ -5,7 +5,7 @@ import better.scoreboard.core.animation.Animation;
 import better.scoreboard.core.bridge.ConfigSection;
 import org.jetbrains.annotations.Nullable;
 
-public class HealthAnimation extends Animation<Float> {
+public class HealthAnimation extends Animation<Double> {
 
     /**
      * Initialize the HealthAnimation, reading required data from the configuration.
@@ -14,11 +14,11 @@ public class HealthAnimation extends Animation<Float> {
         super(plugin, config);
 
         if (config == null) {
-            animation.add(1.0F);
+            animation.add(1.0);
             return;
         }
 
-        animation.addAll(config.getList(Float.class, "health"));
+        animation.addAll(config.getList(Double.class, "health"));
         if (random) currentIndex = (int) (animation.size() * Math.random());
     }
 }
